requirements.txt
streamlit>=1.52.0
pandas>=2.3.0
openpyxl>=3.1.5
textblob>=0.19.0
matplotlib>=3.10.0
altair>=5.5.0
google-generativeai
